requirements.txt
Flask==3.1.1
openai==1.96.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
cloudinary==1.44.1
Pillow==11.0.0
Werkzeug==3.1.3
bcrypt==4.3.0